Rename gateway_dataclasses.py to interface.py
